requirements.txt
aiogram_dialog==2.1.0
aiogram==3.6.0
python_dotenv==1.0.1